requirements.txt
Flask==1.0.2
psycopg2==2.7.5
PyJWT==1.6.4
pytest-cov==2.6.0
flasgger==0.8.0